requirements.txt
feedparser==6.0.12
requests==2.32.5
Flask==3.1.3
python-dateutil==2.9.0.post0
